+++ parse_examples/function/function_unnest.rs
@@ -2,6 +2,6 @@ fn main() {
     another_function(5);
 }
 
-fn another_function(x: i32) {
+fn another_function(x: i64) {
     println("The value of x is: {}", x);
 }
