+++ parse_examples/weird_rust/borrow_method_error.rs
@@ -7,7 +7,7 @@ fn main() {
     println("iis it 5? : {}", result);
 }
 
-fn can_change(alt_ptr_to_content : &mut i32, ptr_to_ptr : &mut &mut i32) -> i32 {
+fn can_change(alt_ptr_to_content : &mut i64, ptr_to_ptr : &mut &mut i64) -> i64 {
     *alt_ptr_to_content = 5;
     **ptr_to_ptr = 13;
     return *alt_ptr_to_content;
